# copilot/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookYourCoach API",
        "description": "API REST pour la plateforme de réservation de cours avec coaches équestres ou autres sports. Cette API permet la gestion complète des utilisateurs (admin/enseignants/élèves), des réservations, des paiements et de la facturation.",
        "contact": {
            "name": "BookYourCoach Support",
            "url": "https://bookyourcoach.com",
            "email": "support@bookyourcoach.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Serveur de développement BookYourCoach"
        }
    ],
    "paths": {
...
        "/api/course-types": {
            "get": {
                "tags": [
                    "Course Types"
                ],
                "summary": "Liste des types de cours",
                "description": "Récupère la liste complète des types de cours disponibles",
                "operationId": "getCourseTypes",
                "responses": {
                    "200": {
                        "description": "Liste des types de cours récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CourseType"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Course Types"
                ],
                "summary": "Créer un nouveau type de cours",
                "description": "Crée un nouveau type de cours",
                "operationId": "createCourseType",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dressage"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Cours de dressage classique"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type de cours créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseType"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Type de cours créé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/course-types/{id}": {
            "get": {
                "tags": [
                    "Course Types"
                ],
                "summary": "Détails d'un type de cours",
                "description": "Récupère les détails d'un type de cours spécifique",
                "operationId": "getCourseType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type de cours",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du type de cours récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Type de cours non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Course Types"
                ],
                "summary": "Mettre à jour un type de cours",
                "description": "Met à jour les informations d'un type de cours",
                "operationId": "updateCourseType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type de cours",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dressage Avancé"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Cours de dressage niveau avancé"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 90
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 65
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type de cours mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseType"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Type de cours mis à jour avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Type de cours non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Course Types"
                ],
                "summary": "Supprimer un type de cours",
                "description": "Supprime un type de cours",
                "operationId": "deleteCourseType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type de cours",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type de cours supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Type de cours supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Type de cours non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CourseType": {
                "title": "Course Type",
                "description": "Type de cours disponible sur la plateforme",
                "properties": {
                    "id": {
                        "description": "Identifiant unique du type de cours",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom du type de cours",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Dressage"
                    },
                    "description": {
                        "description": "Description du type de cours",
                        "type": "string",
                        "example": "Cours de dressage classique"
                    },
                    "duration": {
                        "description": "Durée standard en minutes",
                        "type": "integer",
                        "example": 60
                    },
                    "price": {
                        "description": "Prix de base",
                        "type": "number",
                        "format": "float",
                        "example": 45
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour",
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Bearer. Utilisez le token obtenu via /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}